+++ R/scale_color_bluebrown.R
@@ -16,13 +16,13 @@
 #'   theme_modern() +
 #'   scale_fill_bluebrown_d()
 #' @export
-scale_color_bluebrown <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
+scale_color_bluebrown <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
   pal <- palette_bluebrown(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("bluebrown_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("bluebrown_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -33,14 +33,14 @@ scale_color_bluebrown <- function(palette = "contrast", discrete = TRUE, reverse
 
 #' @rdname scale_color_bluebrown
 #' @export
-scale_color_bluebrown_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
-  scale_color_bluebrown(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_bluebrown_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_bluebrown(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_bluebrown
 #' @export
-scale_color_bluebrown_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, ...) {
-  scale_color_bluebrown(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_bluebrown_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_bluebrown(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_bluebrown
@@ -65,27 +65,27 @@ scale_colour_bluebrown_d <- scale_color_bluebrown_d
 
 #' @rdname scale_color_bluebrown
 #' @export
-scale_fill_bluebrown <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
+scale_fill_bluebrown <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
   pal <- palette_bluebrown(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("bluebrown_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("bluebrown_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
 
 #' @rdname scale_color_bluebrown
 #' @export
-scale_fill_bluebrown_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
-  scale_fill_bluebrown(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_bluebrown_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_bluebrown(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_bluebrown
 #' @export
-scale_fill_bluebrown_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, ...) {
-  scale_fill_bluebrown(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_bluebrown_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_bluebrown(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 
@@ -110,7 +110,7 @@ bluebrown_colors_list <- c(
 
 #' Extract blue-brown colors as hex codes
 #'
-#' Can be used to get the hex code of specific colors from the blue-brown color palette. Use `bluebrown_colors()` to see all available color.
+#' Can be used to get the hex code of specific colors from the blue-brown color palette. Use `bluebrown_colors()` to see all available colors.
 #'
 #' @inheritParams flat_colors
 #'
@@ -156,9 +156,5 @@ bluebrown_palettes <- list(
 #'
 #' @export
 palette_bluebrown <- function(palette = "contrast", reverse = FALSE, ...) {
-  pal <- bluebrown_palettes[[palette]]
-
-  if (reverse) pal <- rev(pal)
-
-  grDevices::colorRampPalette(pal, ...)
+  .retrieve_palette(palette, bluebrown_palettes, reverse = reverse, ...)
 }