+++ R/scale_color_see.R
@@ -28,13 +28,14 @@
 scale_color_see <- function(palette = "contrast",
                             discrete = TRUE,
                             reverse = FALSE,
+                            aesthetics = "color",
                             ...) {
   pal <- palette_see(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("see_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("see_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -48,11 +49,13 @@ scale_color_see <- function(palette = "contrast",
 scale_color_see_d <- function(palette = "contrast",
                               discrete = TRUE,
                               reverse = FALSE,
+                              aesthetics = "color",
                               ...) {
   scale_color_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -62,11 +65,13 @@ scale_color_see_d <- function(palette = "contrast",
 scale_color_see_c <- function(palette = "contrast",
                               discrete = FALSE,
                               reverse = FALSE,
+                              aesthetics = "color",
                               ...) {
   scale_color_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -96,13 +101,14 @@ scale_colour_see_d <- scale_color_see_d
 scale_fill_see <- function(palette = "contrast",
                            discrete = TRUE,
                            reverse = FALSE,
+                           aesthetics = "fill",
                            ...) {
   pal <- palette_see(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("see_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("see_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -112,11 +118,13 @@ scale_fill_see <- function(palette = "contrast",
 scale_fill_see_d <- function(palette = "contrast",
                              discrete = TRUE,
                              reverse = FALSE,
+                             aesthetics = "fill",
                              ...) {
   scale_fill_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -126,11 +134,13 @@ scale_fill_see_d <- function(palette = "contrast",
 scale_fill_see_c <- function(palette = "contrast",
                              discrete = FALSE,
                              reverse = FALSE,
+                             aesthetics = "fill",
                              ...) {
   scale_fill_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -163,7 +173,7 @@ see_colors_list <- c(
 #' Extract See colors as hex codes
 #'
 #' Can be used to get the hex code of specific colors from the See color
-#' palette. Use `see_colors()` to see all available color.
+#' palette. Use `see_colors()` to see all available colors.
 #'
 #' @inheritParams flat_colors
 #'
@@ -204,7 +214,26 @@ see_palettes <- list(
 #'
 #' @export
 palette_see <- function(palette = "contrast", reverse = FALSE, ...) {
-  pal <- see_palettes[[palette]]
+  .retrieve_palette(palette, see_palettes, reverse = reverse, ...)
+}
+
+
+
+
+
+# helper -----------------------
+
+.retrieve_palette <- function(palette, palette_list, reverse = FALSE, ...) {
+  if (!palette %in% names(palette_list)) {
+    msg <- c(paste0(
+      "Palette name not available. `palette` must be one of ",
+      datawizard::text_concatenate(names(palette_list), last = " or ", enclose = "`"),
+      "."), "Using default palette now."
+    )
+    warning(insight::format_message(msg), call. = FALSE)
+    palette <- 1
+  }
+  pal <- palette_list[[palette]]
 
   if (reverse) pal <- rev(pal)